Implement a (hopefully) more smoother message list
